pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "nts.hardware"
dynamic = ["version"]
description = "Different hardware controls"
authors = [
  { name = "Anton Bondarenko", email = "bond.anton@gmail.com" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.9"
dependencies = [
    "gpiozero",
    "python-statemachine"
]

[project.urls]
"Homepage" = "https://github.com/bond-anton/nts.hardware"
"Bug Tracker" = "https://github.com/bond-anton/nts.hardware/issues"

[project.optional-dependencies]
hw_pwm = ["rpi-hardware-pwm"]
dev = ["pydot", "graphviz"]
test = ["flake8", "pytest"]
lint = ["pylint"]

[tool.setuptools.dynamic]
version = {attr = "nts.hardware.__version__"}

[tool.mypy]
python_version = "3.9"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
